refactor pkg client
