--- pkg/adapters/clientv2_5_0/go.mod
@@ -1,14 +1,0 @@
-module clientv2_5_0
-
-go 1.13
-
-require (
-	github.com/vesoft-inc/nebula-go/v2 v2.5.0
-	github.com/vesoft-inc/nebula-http-gateway/pkg/logger v0.0.0
-	github.com/vesoft-inc/nebula-http-gateway/pkg/types v0.0.0
-)
-
-replace (
-	github.com/vesoft-inc/nebula-http-gateway/pkg/logger v0.0.0 => ../../logger
-	github.com/vesoft-inc/nebula-http-gateway/pkg/types v0.0.0 => ../../types
-)
