+++ pkg/client/clientv2_5_0/session.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	nebula "github.com/vesoft-inc/nebula-go/v2"
-	"github.com/vesoft-inc/nebula-http-gateway/pkg/types"
+	"github.com/vesoft-inc/nebula-http-gateway/pkg/client/types"
 )
 
 type Session struct {
