+++ pkg/client/clientv2_5_1/wrapper.go
@@ -1,9 +1,9 @@
-package clientv2_6_0
+package clientv2_5_1
 
 import (
 	"errors"
 	nebula "github.com/vesoft-inc/nebula-go/v2"
-	"github.com/vesoft-inc/nebula-http-gateway/pkg/types"
+	"github.com/vesoft-inc/nebula-http-gateway/pkg/client/types"
 )
 
 func configWrapper(config types.ClientConfig) nebula.PoolConfig {
