+++ pkg/client/clientv2_6_0/session.go
@@ -1,10 +1,8 @@
-package clientv2_5_1
+package clientv2_6_0
 
 import (
-	"errors"
-
 	nebula "github.com/vesoft-inc/nebula-go/v2"
-	"github.com/vesoft-inc/nebula-http-gateway/pkg/types"
+	"github.com/vesoft-inc/nebula-http-gateway/pkg/client/types"
 )
 
 type Session struct {
@@ -24,5 +22,9 @@ func (s *Session) Execute(stmt string) (types.ResultSet, error) {
 }
 
 func (s *Session) ExecuteJson(stmt string) ([]byte, error) {
-	return nil, errors.New("execute json not support")
+	resultAsBytes, err := s.session.ExecuteJson(stmt)
+	if err != nil {
+		return nil, err
+	}
+	return resultAsBytes, nil
 }
