+++ pkg/client/logger/go.mod
@@ -1,1 +1,3 @@
 module logger
+
+go 1.13
