+++ pkg/client/types/go.mod
@@ -1,1 +1,3 @@
 module types
+
+go 1.13
