+++ pkg/client/types/session.go
